refactor: added 'impl_eth_data_provider' macro
